+++ server/src/code-lens.ts
@@ -1,64 +1,57 @@
-import {Diagnostic, DiagnosticSeverity} from 'vscode-languageserver'
+import {CodeLens} from 'vscode-languageserver'
 import {TextDocument} from 'vscode-languageserver-textdocument'
 import {CSSServiceMap, HTMLServiceMap, PartType} from './languages'
 import {getPathExtension} from './utils'
 import {CSSSelectorDetailedPart} from './languages/parts/part-css-selector-detailed'
 
 
-/** Provide class name diagnostics service. */
-export async function getDiagnostics(
+/** Provide class name CodeLens service. */
+export async function getCodeLens(
 	document: TextDocument,
 	htmlServiceMap: HTMLServiceMap,
 	cssServiceMap: CSSServiceMap,
 	configuration: Configuration
-): Promise<Diagnostic[] | null> {
+): Promise<CodeLens[] | null> {
 	let documentExtension = getPathExtension(document.uri)
 	let isHTMLFile = configuration.activeHTMLFileExtensions.includes(documentExtension)
 	let isCSSFile = configuration.activeCSSFileExtensions.includes(documentExtension)
-	let shouldProvideDefDiag = isHTMLFile && configuration.enableClassNameDefinitionDiagnostic
-	let shouldProvideRefDiag = (isHTMLFile || isCSSFile) && configuration.enableClassNameReferenceDiagnostic
+	let codeLens: CodeLens[] = []
 
-	if (!shouldProvideDefDiag && !shouldProvideRefDiag) {
-		return null
-	}
-
-	let diagnostics: Diagnostic[] = []
-
-	if (shouldProvideDefDiag) {
-		let diags = await getDefinitionDiagnostics(document, htmlServiceMap, cssServiceMap, configuration)
+	if (isHTMLFile && configuration.enableDefinitionCodeLens) {
+		let diags = await getDefinitionCodeLens(document, htmlServiceMap, cssServiceMap, configuration)
 		if (diags) {
-			diagnostics.push(...diags)
+			codeLens.push(...diags)
 		}
 	}
 
-	if (shouldProvideRefDiag) {
-		let diags = await getReferencedDiagnostics(document, htmlServiceMap, cssServiceMap, configuration)
+	if ((isHTMLFile || isCSSFile) && configuration.enableReferenceCodeLens) {
+		let diags = await getReferencedCodeLens(document, htmlServiceMap, cssServiceMap, configuration)
 		if (diags) {
-			diagnostics.push(...diags)
+			codeLens.push(...diags)
 		}
 	}
 
-	return diagnostics
+	return codeLens
 }
 
 
-/** Provide defined class name diagnostics service. */
-async function getDefinitionDiagnostics(
+/** Provide defined class name code lens service. */
+async function getDefinitionCodeLens(
 	document: TextDocument,
 	htmlServiceMap: HTMLServiceMap,
 	cssServiceMap: CSSServiceMap,
 	configuration: Configuration
-): Promise<Diagnostic[] | null> {
+): Promise<CodeLens[] | null> {
 	let currentHTMLService = await htmlServiceMap.forceGetServiceByDocument(document)
 	if (!currentHTMLService) {
 		return null
 	}
 
-	let diagnostics: Diagnostic[] = []
+	let codeLens: CodeLens[] = []
 
 	let classNameParts = currentHTMLService.getPartsByType(PartType.Class)
 	if (!classNameParts || classNameParts.length === 0) {
-		return diagnostics
+		return codeLens
 	}
 
 	await cssServiceMap.beFresh()
@@ -71,45 +64,45 @@ async function getDefinitionDiagnostics(
 
 		// Without identifier.
 		let className = part.text
+		let count = 0
 
-		if (currentHTMLService.hasDefinedClassName(className)) {
-			continue
-		}
-
-		if (cssServiceMap.hasDefinedClassName(className)) {
-			continue
-		}
+		count += cssServiceMap.getDefinedClassNameCount(className)
 
 		if (configuration.enableGlobalEmbeddedCSS) {
-			if (htmlServiceMap.hasDefinedClassName(className)) {
-				continue
-			}
+			count += htmlServiceMap.getDefinedClassName(className)
+		}
+		else {
+			count += currentHTMLService.getDefinedClassNameCount(className)
 		}
 
-		diagnostics.push({
-			severity: DiagnosticSeverity.Warning,
-            range: {start: document.positionAt(part.start), end: document.positionAt(part.end)},
-            message: `Can't find definition for ".${className}".`,
-            source: 'CSS Navigation',
-		})
+		if (count > 0) {
+			codeLens.push({
+				range: {start: document.positionAt(part.start), end: document.positionAt(part.end)},
+				command: {
+					title: count > 1 ? `${count} definitions` : `${count} definition`,
+					command: `CSSNavigation.peekDefinitions`,
+					arguments: [document.uri, document.positionAt(part.start)],
+				},
+			})
+		}
 	}
 
-	return diagnostics
+	return codeLens
 }
 
 
 
-/** Provide referenced class name diagnostics service. */
-async function getReferencedDiagnostics(
+/** Provide referenced class name CodeLens service. */
+async function getReferencedCodeLens(
 	document: TextDocument,
 	htmlServiceMap: HTMLServiceMap,
 	cssServiceMap: CSSServiceMap,
 	configuration: Configuration
-): Promise<Diagnostic[] | null> {
+): Promise<CodeLens[] | null> {
 	let documentExtension = getPathExtension(document.uri)
 	let isHTMLFile = configuration.activeHTMLFileExtensions.includes(documentExtension)
 	let isCSSFile = configuration.activeCSSFileExtensions.includes(documentExtension)
-	let diagnostics: Diagnostic[] = []
+	let codeLens: CodeLens[] = []
 
 	if (isHTMLFile) {
 		let currentHTMLService = await htmlServiceMap.forceGetServiceByDocument(document)
@@ -119,7 +112,7 @@ async function getReferencedDiagnostics(
 
 		let classNameParts = currentHTMLService.getPartsByType(PartType.CSSSelectorClass) as CSSSelectorDetailedPart[] | undefined
 		if (!classNameParts || classNameParts.length === 0) {
-			return diagnostics
+			return codeLens
 		}
 
 		if (configuration.enableGlobalEmbeddedCSS) {
@@ -134,35 +127,34 @@ async function getReferencedDiagnostics(
 			}
 
 			let classNames = part.formatted
+			let count = 0
 
 			for (let className of classNames) {
 
 				// Without identifier.
 				let nonIdentifierClassName = className.slice(1)
 
-				// Find only within current document.
-				// Any one of formatted exist, break.
-				if (currentHTMLService.hasReferencedClassName(nonIdentifierClassName)) {
-					break
-				}
-
 				if (configuration.enableGlobalEmbeddedCSS) {
-					if (htmlServiceMap.hasReferencedClassName(nonIdentifierClassName)) {
-						break
-					}
+					count += htmlServiceMap.getReferencedClassNameCount(nonIdentifierClassName)
 				}
+				else {
+					count += currentHTMLService.getReferencedClassNameCount(nonIdentifierClassName)
+				}
+			}
 
-				diagnostics.push({
-					severity: DiagnosticSeverity.Warning,
+			if (count > 0) {
+				codeLens.push({
 					range: {start: document.positionAt(part.start), end: document.positionAt(part.end)},
-					message: `Can't find reference for "${className}".`,
-					source: 'CSS Navigation',
+					command: {
+						title: count > 1 ? `${count} references` : `${count} reference`,
+						command: `CSSNavigation.peekReferences`,
+						arguments: [document.uri, document.positionAt(part.start)],
+					},
 				})
-				break
 			}
 		}
 
-		return diagnostics
+		return codeLens
 	}
 	else if (isCSSFile) {
 		let currentCSSService = await cssServiceMap.forceGetServiceByDocument(document)
@@ -172,7 +164,7 @@ async function getReferencedDiagnostics(
 
 		let classNameParts = currentCSSService.getPartsByType(PartType.CSSSelectorClass) as CSSSelectorDetailedPart[] | undefined
 		if (!classNameParts || classNameParts.length === 0) {
-			return diagnostics
+			return codeLens
 		}
 
 		await htmlServiceMap.beFresh()
@@ -185,6 +177,7 @@ async function getReferencedDiagnostics(
 			}
 
 			let classNames = part.formatted
+			let count = 0
 
 			for (let className of classNames) {
 
@@ -192,21 +185,22 @@ async function getReferencedDiagnostics(
 				let nonIdentifierClassName = className.slice(1)
 
 				// Any one of formatted exist, break.
-				if (htmlServiceMap.hasReferencedClassName(nonIdentifierClassName)) {
-					break
-				}
+				count += htmlServiceMap.getReferencedClassNameCount(nonIdentifierClassName)
+			}
 
-				diagnostics.push({
-					severity: DiagnosticSeverity.Warning,
+			if (count > 0) {
+				codeLens.push({
 					range: {start: document.positionAt(part.start), end: document.positionAt(part.end)},
-					message: `Can't find reference for "${className}".`,
-					source: 'CSS Navigation',
+					command: {
+						title: count > 1 ? `${count} references` : `${count} reference`,
+						command: `CSSNavigation.peekReferences`,
+						arguments: [document.uri, document.positionAt(part.start)],
+					},
 				})
-				break
 			}
 		}
 
-		return diagnostics
+		return codeLens
 	}
 
 	return null